Public API

YAXArrays.Cubes — Module

The functions provided by YAXArrays are supposed to work on different types of cubes. This module defines the interface for all Data types that

YAXArrays.Cubes.YAXArray — Type

YAXArray{T,N}

An array labelled with named axes that have values associated with them. It can wrap normal arrays or, more typically DiskArrays.

Fields

axes a Vector{CubeAxis} containing the Axes of the Cube

data N-D array containing the data

YAXArrays.Cubes.caxes — Function

Returns the axes of a Cube

YAXArrays.Cubes.concatenatecubes — Method

function concatenateCubes(cubelist, cataxis::CategoricalAxis)

Concatenates a vector of datacubes that have identical axes to a new single cube along the new axis cataxis

YAXArrays.Cubes.readcubedata — Method

readcubedata(cube)

Given any array implementing the YAXArray interface it returns an in-memory YAXArray from it.

YAXArrays.Cubes.setchunks — Method

setchunks(c::YAXArray,chunks)

Resets the chunks of a YAXArray and returns a new YAXArray. Note that this will not change the chunking of the underlying data itself, it will just make the data "look" like it had a different chunking. If you need a persistent on-disk representation of this chunking, use savecube on the resulting array. The chunks argument can take one of the following forms:

a DiskArrays.GridChunks object

a tuple specifying the chunk size along each dimension

an AbstractDict or NamedTuple mapping one or more axis names to chunk sizes

YAXArrays.Cubes.subsetcube — Function

This function calculates a subset of a cube's data

YAXArrays.Cubes.Axes — Module

The Axes module handles the Axes of a data cube. It provides the following exports:

CategoricalAxis

CubeAxis

RangeAxis

getAxis

YAXArrays.Cubes.Axes.CategoricalAxis — Type

struct CategoricalAxis{T,S,RT}

To represent axes that are categorical, where T is the element type. The type parameter S denotes the axis name (a symbol). The default constructor is:

CategoricalAxis(axname::String,values::Vector{T})

YAXArrays.Cubes.Axes.CubeAxis — Type

abstract type CubeAxis{T,S}

Supertype of all axes. Every CubeAxis is an 1D Cube itself and can be passed to mapCube operations. In detail CubeAxis is an AbstractArray{Int, 1}

YAXArrays.Cubes.Axes.RangeAxis — Type

RangeAxis{T,S,R}

To represent axes that are numerical, where T is the element type. The type parameter S denotes the axis name (a symbol) and R the type of the range which is used to represent the axis values. The default constructor is:

RangeAxis(axname::String,values::Range{T})

YAXArrays.Cubes.Axes.getAxis — Method

getAxis(desc, c)

Given an Axis description and a cube, returns the corresponding axis of the cube. The Axis description can be:

the name as a string or symbol.

an Axis object

YAXArrays.DAT.InDims — Type

InDims(axisdesc...;...)

Creates a description of an Input Data Cube for cube operations. Takes a single or multiple axis descriptions as first arguments. Alternatively a MovingWindow(@ref) struct can be passed to include neighbour slices of one or more axes in the computation. Axes can be specified by their name (String), through an Axis type, or by passing a concrete axis.

Keyword arguments

artype how shall the array be represented in the inner function. Defaults to Array, alternatives are DataFrame or AsAxisArray

filter define some filter to skip the computation, e.g. when all values are missing. Defaults to AllMissing(), possible values are AnyMissing(), AnyOcean(), StdZero(), NValid(n) (for at least n non-missing elements). It is also possible to provide a custom one-argument function that takes the array and returns true if the compuation shall be skipped and false otherwise.

window_oob_value if one of the input dimensions is a MowingWindow, this value will be used to fill out-of-bounds areas

YAXArrays.DAT.MovingWindow — Type

MovingWindow(desc, pre, after)

Constructs a MovingWindow object to be passed to an InDims constructor to define that the axis in desc shall participate in the inner function (i.e. shall be looped over), but inside the inner function pre values before and after values after the center value will be passed as well.

For example passing MovingWindow("Time", 2, 0) will loop over the time axis and always pass the current time step plus the 2 previous steps. So in the inner function the array will have an additional dimension of size 3.

YAXArrays.DAT.OutDims — Method

OutDims(axisdesc;...)

Creates a description of an Output Data Cube for cube operations. Takes a single or a Vector/Tuple of axes as first argument. Axes can be specified by their name (String), through an Axis type, or by passing a concrete axis.

axisdesc: List of input axis names

backend : specifies the dataset backend to write data to, must be either :auto or a key in YAXArrayBase.backendlist

update : specifies wether the function operates inplace or if an output is returned

artype : specifies the Array type inside the inner function that is mapped over

chunksize: A Dict specifying the chunksizes for the output dimensions of the cube, or :input to copy chunksizes from input cube axes or :max to not chunk the inner dimensions

outtype: force the output type to a specific type, defaults to Any which means that the element type of the first input cube is used

YAXArrays.DAT.CubeTable — Method

CubeTable()

Function to turn a DataCube object into an iterable table. Takes a list of as arguments, specified as a name=cube expression. For example CubeTable(data=cube1,country=cube2) would generate a Table with the entries data and country, where data contains the values of cube1 and country the values of cube2. The cubes are matched and broadcasted along their axes like in mapCube.

YAXArrays.DAT.cubefittable — Method

cubefittable(tab,o,fitsym;post=getpostfunction(o),kwargs...)

Executes fittable on the CubeTable tab with the (Weighted-)OnlineStat o, looping through the values specified by fitsym. Finally, writes the results from the TableAggregator to an output data cube.

YAXArrays.DAT.fittable — Method

fittable(tab,o,fitsym;by=(),weight=nothing)

Loops through an iterable table tab and thereby fitting an OnlineStat o with the values specified through fitsym. Optionally one can specify a field (or tuple) to group by. Any groupby specifier can either be a symbol denoting the entry to group by or an anynymous function calculating the group from a table row.

For example the following would caluclate a weighted mean over a cube weighted by grid cell area and grouped by country and month:

fittable(iter,WeightedMean,:tair,weight=(i->abs(cosd(i.lat))),by=(i->month(i.time),:country))

YAXArrays.DAT.mapCube — Method

mapCube(fun, cube, addargs...;kwargs...)

Map a given function fun over slices of the data cube cube. The additional arguments addargs will be forwarded to the inner function fun.

Keyword arguments

max_cache=YAXDefaults.max_cache maximum size of blocks that are read into memory, defaults to approx 10Mb

indims::InDims List of input cube descriptors of type InDims for each input data cube

outdims::OutDims List of output cube descriptors of type OutDims for each output cube

inplace does the function write to an output array inplace or return a single value> defaults to true

ispar boolean to determine if parallelisation should be applied, defaults to true if workers are available.

showprog boolean indicating if a ProgressMeter shall be shown

include_loopvars boolean to indicate if the varoables looped over should be added as function arguments

nthreads number of threads for the computation, defaults to Threads.nthreads for every worker.

loopchunksize determines the chunk sizes of variables which are looped over, a dict

kwargs additional keyword arguments are passed to the inner function

The first argument is always the function to be applied, the second is the input cube or a tuple of input cubes if needed.

YAXArrays.Datasets.Dataset — Method

Dataset(; properties = Dict{String,Any}, cubes...)

Construct a YAXArray Dataset with global attributes properties a and a list of named YAXArrays cubes...

YAXArrays.Datasets.Cube — Method

Cube(ds::Dataset; joinname="Variable")

Construct a single YAXArray from the dataset ds by concatenating the cubes in the datset on the joinname dimension.

YAXArrays.Datasets.savecube — Method

savecube(cube,name::String)

Save a YAXArray to the path.

Extended Help

The keyword arguments are:

name:

datasetaxis="Variable" special treatment of a categorical axis that gets written into separate zarr arrays

max_cache: The number of bits that are used as cache for the data handling.

backend: The backend, that is used to save the data. Fallsback to searching the backend according to the extension of the path.

driver

overwrite::Bool=false overwrite cube if it already exists

YAXArrays.Datasets.to_dataset — Method

to_dataset(c;datasetaxis = "Variable", name = "layer")

Convert a Data Cube into a Dataset. It is possible to treat one of the Cube's axes as a "DatasetAxis" i.e. the cube will be split into different parts that become variables in the Dataset. If no such axis is specified or found, there will only be a single variable in the dataset with the name name

Internal API

YAXArrays.Cubes.CleanMe — Type

mutable struct CleanMe

Struct which describes data paths and their persistency. Non-persistend paths/files are removed at finalize step

YAXArrays.Cubes.clean — Method

clean(c::CleanMe)

finalizer function for CleanMe struct. The main process removes all directories/files which are not persistent.

YAXArrays.Cubes.copydata — Method

copydata(outar, inar, copybuf)

Internal function which copies the data from the input inar into the output outar at the copybuf positions.

YAXArrays.Cubes.optifunc — Method

optifunc(s, maxbuf, incs, outcs, insize, outsize, writefac)

Internal

This function is going to be minimized to detect the best possible chunk setting for the rechunking of the data.

YAXArrays.Cubes.Axes.abshalf — Method

abshalf

YAXArrays.Cubes.Axes.axVal2Index — Method

axVal2Index

YAXArrays.Cubes.Axes.axVal2Index_lb — Method

axVal2Index_lb

YAXArrays.Cubes.Axes.axVal2Index_ub — Method

axVal2Index_ub

YAXArrays.Cubes.Axes.axcopy — Method

axcopy(x,vals)

Makes a full copy of a CubeAxis with the values vals

YAXArrays.Cubes.Axes.axisfrombb — Method

axisfrombb

YAXArrays.Cubes.Axes.axname — Method

axname

YAXArrays.Cubes.Axes.axsym — Method

axsym

YAXArrays.Cubes.Axes.convert_time — Method

convert_time

YAXArrays.Cubes.Axes.findAxis — Method

findAxis(desc, c)

Given an Axis description and a cube return the index of the Axis. The Axis description can be:

the name as a string or symbol.

an Axis object

YAXArrays.Cubes.Axes.getOutAxis — Method

getOutAxis

YAXArrays.Cubes.Axes.get_bb — Method

get_bb

YAXArrays.Cubes.Axes.get_descriptor — Method

get_descriptor(a)

Get the descriptor of an Axis. This is used to dispatch on the descriptor.

YAXArrays.Cubes.Axes.get_step — Method

get_step

returns stepwidth of the RangeAxis

YAXArrays.Cubes.Axes.match_axis — Method

match_axis

YAXArrays.Cubes.Axes.renameaxis — Method

renameaxis

YAXArrays.Cubes.caxes — Method

caxes

Embeds Cube inside a new Cube

YAXArrays.DAT.DATConfig — Type

Configuration object of a DAT process. This holds all necessary information to perform the calculations. It contains the following fields:

incubes::NTuple{NIN,InputCube} The input data cubes

outcube::NTuple{NOUT,OutputCube} The output data cubes

allInAxes :: Vector LoopAxes :: Vector ispar :: Bool loopcachesize :: Vector{Int} maxcache fu inplace :: Bool includeloopvars:: Bool ntr addargs kwargs

YAXArrays.DAT.InputCube — Type

Internal representation of an input cube for DAT operations

cube

The input data

desc

The input description given by the user/registration

axesSmall

List of axes that were actually selected through the description

icolon

colonperm

loopinds

Indices of loop axes that this cube does not contain, i.e. broadcasts

cachesize

window

iwindow

windowloopinds

iall

YAXArrays.DAT.OutputCube — Type

Internal representation of an output cube for DAT operations

Fields

cube

The actual outcube cube, once it is generated

cube_unpermuted

The unpermuted output cube

desc

The description of the output axes as given by users or registration

axesSmall

The list of output axes determined through the description

allAxes

List of all the axes of the cube

loopinds

Index of the loop axes that are broadcasted for this output cube

innerchunks

outtype

Elementtype of the outputcube

YAXArrays.DAT.YAXColumn — Type

YAXColumn

A struct representing a single column of a YAXArray partitioned Table # Fields

inarBC

inds

YAXArrays.DAT.cmpcachmisses — Method

Function that compares two cache miss specifiers by their importance

YAXArrays.DAT.getFrontPerm — Method

Calculate an axis permutation that brings the wanted dimensions to the front

YAXArrays.DAT.getLoopCacheSize — Method

Calculate optimal Cache size to DAT operation

YAXArrays.Cubes.setchunks — Method

setchunks(c::Dataset,chunks)

Resets the chunks of all or a subset YAXArrays in the dataset and returns a new Dataset. Note that this will not change the chunking of the underlying data itself, it will just make the data "look" like it had a different chunking. If you need a persistent on-disk representation of this chunking, use savedataset on the resulting array. The chunks argument can take one of the following forms:

a NamedTuple or AbstractDict mapping from variable name to a description of the desired variable chunks

a NamedTuple or AbstractDict mapping from dimension name to a description of the desired variable chunks

a description of the desired variable chunks applied to all members of the Dataset

where a description of the desired variable chunks can take one of the following forms:

a DiskArrays.GridChunks object

a tuple specifying the chunk size along each dimension

an AbstractDict or NamedTuple mapping one or more axis names to chunk sizes

YAXArrays.Datasets.collectfromhandle — Method

Extracts a YAXArray from a dataset handle that was just created from a arrayinfo

YAXArrays.Datasets.createdataset — Method

function createdataset(DS::Type,axlist; kwargs...)

Creates a new dataset with axes specified in axlist. Each axis must be a subtype of CubeAxis. A new empty Zarr array will be created and can serve as a sink for mapCube operations.

Keyword arguments

path="" location where the new cube is stored

T=Union{Float32,Missing} data type of the target cube

chunksize = ntuple(i->length(axlist[i]),length(axlist)) chunk sizes of the array

chunkoffset = ntuple(i->0,length(axlist)) offsets of the chunks

persist::Bool=true shall the disk data be garbage-collected when the cube goes out of scope?

overwrite::Bool=false overwrite cube if it already exists

properties=Dict{String,Any}() additional cube properties

fillvalue= T>:Missing ? defaultfillval(Base.nonmissingtype(T)) : nothing fill value

datasetaxis="Variable" special treatment of a categorical axis that gets written into separate zarr arrays

YAXArrays.Datasets.getarrayinfo — Method

Extract necessary information to create a YAXArrayBase dataset from a name and YAXArray pair

YAXArrays.Datasets.testrange — Method

Test if data in x can be approximated by a step range