Public API

YAXArrays.Cubes — Module

The functions provided by YAXArrays are supposed to work on different types of cubes. This module defines the interface for all Data types that

YAXArrays.Cubes.YAXArray — Type

YAXArray{T,N}

An array labelled with named axes that have values associated with them. It can wrap normal arrays or, more typically DiskArrays.

Fields

axes a Vector{CubeAxis} containing the Axes of the Cube

data N-D array containing the data

YAXArrays.Cubes.caxes — Function

Returns the axes of a Cube

YAXArrays.Cubes.concatenatecubes — Method

function concatenateCubes(cubelist, cataxis::CategoricalAxis)

Concatenates a vector of datacubes that have identical axes to a new single cube along the new axis cataxis

YAXArrays.Cubes.readcubedata — Method

readcubedata(cube)

Given any array implementing the YAXArray interface it returns an in-memory YAXArray from it.

YAXArrays.Cubes.subsetcube — Function

This function calculates a subset of a cube's data

YAXArrays.Cubes.Axes.CategoricalAxis — Type

CategoricalAxis{T,S}

To represent axes that are categorical, where T is the element type. The type parameter S denotes the axis name (a symbol). The default constructor is:

CategoricalAxis(axname::String,values::Vector{T})

YAXArrays.Cubes.Axes.CubeAxis — Type

abstract CubeAxis{T}

Supertype of all axes. Every CubeAxis is an 1D Cube itself and can be passed to mapCube operations.

YAXArrays.Cubes.Axes.RangeAxis — Method

RangeAxis{T,S,R}

To represent axes that are categorical, where T is the element type. The type parameter S denotes the axis name (a symbol) and R the type of the range which is used to represent the axis values. The default constructor is:

RangeAxis(axname::String,values::Range{T})

YAXArrays.Cubes.Axes.getAxis — Method

getAxis(desc::String, c)

Given the string of an axis name and a cube, returns this axis of the cube.

YAXArrays.DAT.InDims — Type

InDims(axisdesc...;...)

Creates a description of an Input Data Cube for cube operations. Takes a single or multiple axis descriptions as first arguments. Alternatively a MovingWindow(@ref) struct can be passed to include neighbour slices of one or more axes in the computation. Axes can be specified by their name (String), through an Axis type, or by passing a concrete axis.

Keyword arguments

artype how shall the array be represented in the inner function. Defaults to Array, alternatives are DataFrame or AsAxisArray

filter define some filter to skip the computation, e.g. when all values are missing. Defaults to AllMissing(), possible values are AnyMissing(), AnyOcean(), StdZero(), NValid(n) (for at least n non-missing elements). It is also possible to provide a custom one-argument function that takes the array and returns true if the compuation shall be skipped and false otherwise.

window_oob_value if one of the input dimensions is a MowingWindow, this value will be used to fill out-of-bounds areas

YAXArrays.DAT.MovingWindow — Type

MovingWindow(desc, pre, after)

Constructs a MovingWindow object to be passed to an InDims constructor to define that the axis in desc shall participate in the inner function (i.e. shall be looped over), but inside the inner function pre values before and after values after the center value will be passed as well.

For example passing MovingWindow("Time", 2, 0) will loop over the time axis and always pass the current time step plus the 2 previous steps. So in the inner function the array will have an additional dimension of size 3.

YAXArrays.DAT.OutDims — Method

OutDims(axisdesc;...)

Creates a description of an Output Data Cube for cube operations. Takes a single or a Vector/Tuple of axes as first argument. Axes can be specified by their name (String), through an Axis type, or by passing a concrete axis.

axisdesc: List of input axis names

backend : specifies the dataset backend to write data to, must be either :auto or a key in YAXArrayBase.backendlist

update : specifies wether the function operates inplace or if an output is returned

artype : specifies the Array type inside the inner function that is mapped over

chunksize: A Dict specifying the chunksizes for the output dimensions of the cube, or :input to copy chunksizes from input cube axes or :max to not chunk the inner dimensions

outtype: force the output type to a specific type, defaults to Any which means that the element type of the first input cube is used

YAXArrays.DAT.CubeTable — Method

CubeTable()

Function to turn a DataCube object into an iterable table. Takes a list of as arguments, specified as a name=cube expression. For example CubeTable(data=cube1,country=cube2) would generate a Table with the entries data and country, where data contains the values of cube1 and country the values of cube2. The cubes are matched and broadcasted along their axes like in mapCube.

In addition, one can specify include_axes=(ax1,ax2...) when one wants to include the values of certain axes in the table. For example the command (CubeTable(tair=cube1 axes=("lon","lat","time")) would produce an iterator over a data structure with entries tair, lon, lat and time.

YAXArrays.DAT.cubefittable — Method

cubefittable(tab,o,fitsym;post=getpostfunction(o),kwargs...)

Executes fittable on the @CubeTable tab with the (Weighted-)OnlineStat o, looping through the values specified by fitsym. Finally, writes the results from the TableAggregator to an output data cube.

YAXArrays.DAT.fittable — Method

fittable(tab,o,fitsym;by=(),weight=nothing)

Loops through an iterable table tab and thereby fitting an OnlineStat o with the values specified through fitsym. Optionally one can specify a field (or tuple) to group by. Any groupby specifier can either be a symbol denoting the entry to group by or an anynymous function calculating the group from a table row.

For example the following would caluclate a weighted mean over a cube weighted by grid cell area and grouped by country and month:

fittable(iter,WeightedMean,:tair,weight=(i->abs(cosd(i.lat))),by=(i->month(i.time),:country))

YAXArrays.DAT.mapCube — Method

mapCube(fun, cube, addargs...;kwargs)

Map a given function fun over slices of the data cube cube.

Keyword arguments

max_cache=1e7 maximum size of blocks that are read into memory, defaults to approx 10Mb

indims::InDims List of input cube descriptors of type InDims for each input data cube

outdims::OutDims List of output cube descriptors of type OutDims for each output cube

inplace does the function write to an output array inplace or return a single value> defaults to true

ispar boolean to determine if parallelisation should be applied, defaults to true if workers are available.

showprog boolean indicating if a ProgressMeter shall be shown

include_loopvars boolean to indicate if the varoables looped over should be added as function arguments

loopchunksize determines the chunk sizes of variables which are looped over, a dict

kwargs additional keyword arguments passed to the inner function

The first argument is always the function to be applied, the second is the input cube or a tuple input cubes if needed.

YAXArrays.DAT.savecube — Method

savecube(cube,name::String)

Save a YAXArray to the folder name in the YAXArray working directory.

YAXArrays.DAT.SentinelMissings.as_sentinel — Method

as_sentinel(x, v)

Reinterprets a Number Array or a Number x so that values in x that equal v will be treated as missing. This is done by reinterpreting the array as a SentinelMissing without copying the data.

YAXArrays.YAXTools.@loadOrGenerate — Macro

macro loadOrGenerate(x...,expression)

Takes a list of variablename=>"Storage Name" pairs. Checks if all datasets can be found on disk and loads them. If not, the datasets will be regenerated by evaluating the given expression.

To force recalculation, call YAXArrays.recalculate(true) before evaluating the macro.

Example

The following lines will check if cubes with the names "Filled" and "Normalized" exist on disk, load them and assign the variable names cube_filled and cube_norm. If the datasets to not exist on disk, they are generated and saved under the given names.

    @loadOrGenerate cube_filled=>"Filled" cube_norm=>"Normalized" begin
    cube_norm   = mapCube(normalize_TS,d)
end  

Internal API

YAXArrays.Cubes.Axes.axcopy — Method

axcopy(x,vals)

Makes a copy of a CubeAxis with the values vals

YAXArrays.Cubes.Axes.findAxis — Method

Fallback method

YAXArrays.DAT.DATConfig — Type

Configuration object of a DAT process. This holds all necessary information to perform the calculations. It contains the following fields:

incubes::NTuple{NIN,InputCube} The input data cubes

outcube::NTuple{NOUT,OutputCube} The output data cubes

allInAxes :: Vector LoopAxes :: Vector ispar :: Bool loopcachesize :: Vector{Int} maxcache fu inplace :: Bool includeloopvars:: Bool ntr addargs kwargs

YAXArrays.DAT.InputCube — Type

Internal representation of an input cube for DAT operations

YAXArrays.DAT.OutputCube — Type

Internal representation of an output cube for DAT operations

YAXArrays.DAT.cmpcachmisses — Method

Function that compares two cache miss specifiers by their importance

YAXArrays.DAT.getFrontPerm — Method

Calculate an axis permutation that brings the wanted dimensions to the front

YAXArrays.DAT.getLoopCacheSize — Method

Calculate optimal Cache size to DAT operation

YAXArrays.Datasets.createdataset — Method

function createdataset(DS::Type,axlist; kwargs...)

Creates a new datacube with axes specified in axlist. Each axis must be a subtype of CubeAxis. A new empty Zarr array will be created and can serve as a sink for mapCube operations.

Keyword arguments

folder=tempname() location where the new cube is stored

T=Union{Float32,Missing} data type of the target cube

chunksize = ntuple(i->length(axlist[i]),length(axlist)) chunk sizes of the array

chunkoffset = ntuple(i->0,length(axlist)) offsets of the chunks

persist::Bool=true shall the disk data be garbage-collected when the cube goes out of scope?

overwrite::Bool=false overwrite cube if it already exists

properties=Dict{String,Any}() additional cube properties

fillvalue= T>:Missing ? defaultfillval(Base.nonmissingtype(T)) : nothing fill value

datasetaxis="Variable" special treatment of a categorical axis that gets written into separate zarr arrays

YAXArrays.Datasets.testrange — Method

Test if data in x can be approximated by a step range